Fix chunk positioning
Center the chunk in the middle, align the y-axis with the cube edge.

Once I render a 16x16x16 chunk I might need to change the y-alignment.

I don't know if I want it also centered, or aligned at the top or
bottom. I think top or bottom but not sure.