Use tuples as keys instead of strings
This was way harder than it should have been. But I mean, I kind of get
it now. There is now standard support for hashing a tuple, so I
basically had to add it. Makes sense at a high level.

The code I copied though, that makes no sense. I have not learned C++
templates and jeez the syntax in that header file is rough from a
beginner's glance.

For now, it's enough that it works.