Add Chunk destructor to print when a Chunk is destroyed.
This is pretty useful because as it turns out, I'm doing a ton of
destructions. I think I might be copying the Chunk when I pull it out of
the map and then that copy gets destroyed at the end of the scope.

I'll have to look into this more.