Add copy constructor to print when it gets called.
I suspected that the copy constructor was being called when I saw the
stream of destruction prints. I also added a print to the normal
constructor which did not show up, which led me to the hunch that the
copy constructor was being called.

Now that I have a better idea of what's going on, I was able to fix
this, which I'll commit separately.