Don't invoke copy constructor every frame when pulling chunk from map
The previous code was creating a whole new chunk object and throwing it
away each frame because I was accidentally doing a copy initialization
instead of just getting a reference to the existing object in the map.

I still have a few extra copy initializations and destructions at the
start of the program, so I want to fix that next.