Add pointers to debug output
I wanted this so I could see the exact order of what was happening. It
made it clear which object was being copied and destroyed.

I also added the block code to the copy constructor. Since it was still
being called, but didn't set the blocks member, nothing would show up
and it was annoying. So this is a good temporary measure to make
debugging this a bit more obvious.