Generate chunks in the y direction as well
I was wondering if I would want a different viewing distance for the
y-axis. I feel like it should be treated a bit differently.

Hmm maybe not.

The scenario I was imagine was: a large structure that is multiple
chunks high and we want to be able to see the whole thing at once.

I just realized however that my current viewing distance is so tiny, but
in reality it will be much much larger, 8 or 12 chunks maybe. That
should be totally sufficient for viewing the whole thing or at least
enough of it to be sufficient.

So I'm pretty happy with this change and I think it's the way to go.