Move cube vertices, vao, and vbo into Chunk
I'm not sure if this is super efficient, or maybe we have to do this?

Before having a single VAO and a single VBO sufficed, but now we will
have tons of different vertices, so I think we need multiple VAOs to
store them?

Or maybe there's a way to combine them in a single one? No idea, but we
can look into that later.