Create "quads" abstraction
The Greedy Meshing algorithm I'm looking at spits out a collection of
quads, so I wanted to first write the code that would consume that
collection.

This was pretty annoying. In order to handle the face being in any axis,
we have to detect it's orientation -- by checking which axis has
all constant values -- and using that assign the texture coords as
appropriate.