Remove saved comments from Makefile
My initial build overhaul did separate compilation for the various third
party libs.

However, the drawback with how it was structured is that I needed to
specify manually various paths and variable names all specific to each
lib.

The way I do it now just groups everything that needs to be compiled in
third_party_src and has a few rules for c vs cpp files. This should be
more extensible without having to add more to the Makefile later.