Fix Makefile bugs
Need to specify src/* and src/**/* as prereqs to rebuild when anything
in src or src/shaders changes. I don't know if it handles 3 levels or
not, but for now this works.

Also had to set the debug task as a dep of the core and gdb tasks so
that it would rebuild the debug bin if necessary.