Derive cameraFront instead of storing value
cameraFront can be completely determined by yaw and pitch, so deriving
it like this prevents the bug I've had for a while where the default yaw
and pitch might not match the default cameraFront if I've been fiddling
with the starting values (which I did).

This seems less efficient since it has to recompute that vector all the
time, but this doesn't matter now and the code looks nicer.

I could probably have done or could do `computeCameraFront` and use this
value everywhere so that we only change cameraFront when we need to.

I might just do this anyways, I don't know, we'll see.