Simplify firstMouseMovement code
This is maybe too clever, but we're gonna do it anyways.

This takes advantage of the fact that a local static variable is only
ever initialized once.

Therefore, instead of keeping track of the firstMouseMovement bool
ourself, we can actually let the compiler do it.
See: https://stackoverflow.com/a/5567571

This localizes the variable to the one spot it's needed which makes the
class definition look much cleaner. It also makes it clear that these
bookkeeping variables are only ever used here and not for anything else.