Add way to lazy load Textures
This way, we can just use the fetch method and if it's already been
loaded it will get returned, and if not, it will be created.

There are some subtle bugs here I think I need to fix. Right now, if it
has to create it, it will change the current active texture image unit.
But if it's already loaded, it just returns the Texture object.

So the behavior doesn't quite match between the different paths.

I will probably move all functionality that affects the active texture
image unit to the `use` method.