Write first test for MesherGreedy
Now that I am trying to write tests, I'm going to have to deal with the
OpenGL context more intelligently. Running OpenGL lines segfault when
run in a test because there is no context.

So either I need to not run them here, or I need to set up a context
first.

My inclination is to do the former. That is, I should write as much code
as possible that is free from OpenGL as possible.

Something like Shader and Texture, obviously, don't really make sense to
try to split up like that, since they are basically wrappers for OpenGL
methods.

However, when it comes to voxel engine specific files, like Chunk or
MesherGreedy, it seems very plausible that these should be able to work
without any OpenGL code.

I believe decoupling when possible is a "better" design, so I'll
generate the vao and vbo in the render method instead.

---

Also, this commit splits the automatic mesh generation out of the chunk
class. This is a bit of test driven design since I need to be able to
call the meshing method separately instead of having it happen when the
chunk is created in the test. So far, I think I like it. Perhaps Chunk
can even eventually be just a struct. I don't need to do any major
changes now, but I'll keep my mind on it as I keep going on, esp when
writing tests.

---

The next thing our test will require is having a way to tell the chunk
how to generate its blocks. This will solve the whole "commented out
block code in constructor" thing. I don't know what I'll call these
things, idk ChunkGenerator, whatever. I believe I will add another
constructor to Chunk to take a ChunkGenerator instance which will have a
method to generate a blocks array given a position.

I probably won't use the position right away, but I think it's a very
obvious thing needed once we start doing procedural generation which
will required coordinates. I can probably wait to add it until then.