Create simple chunk generator methods
This isn't the final API I like. I'd rather pass something into the
Chunk constructor as an optional parameter. That's kind of required to
be able to write the sort of tests I want to write.

But this moves the code to a more appropriate place, it creates the
files, so I want to commit this for now.