Switch completely over to using a 2d texture array
Now that the texture is encoded in the quad mesh as the array index, we
no longer need to pass around blockType with the vertices or bind the
texture in the loop.

This will enable us to combine all the vertices into a single vao/vbo
instead of iterating over them, which means we can do it on mesh instead
of every frame, which will be more efficient.