Fix shadow mapping not working when not in origin chunk
We need to translate both the light point and the point the light is
aimed at so that the directional rays stay parallel with the x axis.

The shadow texture size is also increased by a bunch in order to give it
more resolution which is needed to reduce shadow acne on large scenes.

Right now, 2^14 is a sufficient size for a viewingDistance of 3.
Eventually, we will need CSM in order to reduce the shadow resolution
and to handle larger scenes.