Fix light angle calculation
Didn't realize this before, but now that we have a 24h time, the entire
range of the angle is 360 degrees whereas before we had only a slider
for angle that went from 0 to 180.

Now, dawn and dusk are at 180 and 0 degrees respectively as they should
be. Midnight is at 270 and noon is at 90.