Introduce proper DirectionalLight concept
With tests! I had experimented with the code in renderer.h long enough
and wanted to have a proper object with methods under test.

The next thing I want to change quickly is to replaced lerp with a bell
curve (I'll probably implement this as an option hooked up to the GUI),
and then after that add a Moon to the scene during night.

Then I also _still_ need to fix the lack of ambient diffuse light. I'm
not quite sure how I was planning on solving that, I definitely had a
few possible solutions, so I'll have to look back through my notes and
maybe discord :grimace: to find what I plan on doing.