Update DirectionalLight to take an offset
So it can be used for both a sun and a moon. The offset passed into the
constructor (in degrees) makes it so that it can provide light at
different times.

I sorta tore apart and rebuilt the DirectionalLight class. One of the
things I got rid of is an angle method used in Renderer. I instead pass
timeOfDay into the ambinentStrength, ambient, diffuse, etc methods
directly.