Change whether sun or moon lighting is used based on time of day
This is for both shadows and lighting. Shadows are a bit tricky (maybe
moreso with forward rendering, reading some stuff that makes it seem
deferred make multiple light sources easier but idk how). Anyways, I
opted to have only a single directional light used for shadows at a
time.

Similarly, the lighting used is only one of the two. However, this I can
change, and pretty easily. All I need to do is add it to the lights
array in the shader so both are added together. That'll be next.