Generate chunk mesh in separate threads
This fixes an issue where crossing chunk boundaries would pause the main
thread for a little bit of time. It was very annoying.

However, with a large enough view distance, there will still be a pause.
I believe this is due to the opengl calls in the Chunk render function
that send the mesh data to the GPU. A more advanced solution might
stagger those calls somehow. Oh, it could also still be that executing
so many threads at once takes away resources from the main (rendering)
thread.