Use a thread pool to remove lag for higher view distances
This is the resolution to the problem I had mentioned in the previous
commit. Now that we use a thread pool, we can simply push all mesh tasks
into it and they will execute as soon as they can.

This removes the overhead of thread creation/deletion, fixes the chunk
render function passing all the new data to the gpu in a single frame,
and fixes having too many threads taking up resources at once.