Add mini-yaml library
Will be (possibly temporarily) used to load block data from yaml config file.

I'm not sure once I add scripting support via mruby if I'll still need
this or not. I really hope not since dealing with this in C++ is kind of
a pain, but I really don't know what using mruby is going to be like.