Add support for blocks with multiple textures
There is now a perf issue when crossing chunk boundaries. I think it's
due to how the chunk generation adds blocks. When I call
Block::createWorldBlock, it loops thru the yaml data for each one to
find the block type which is probably something like O(n) now instead of
O(1) like it was before.

There is another more minor perf issue when the program starts which I
believe is caused by the same exact thing (when it generates the initial
chunks).