Remove most chunk boundary lag
I added a mapping that gets loaded once the first time we need it and
then we pull block type from it ever after. Average access is constant
time and afaict is an appropriate data structure to be used here.

Removes most lag, but there is still a noticeable spike when
crossing boundaries, and I'm not sure why. Needs further investigation.