Move chunk gen off of rendering thread to other threads
This is WIP/MVP code. I'm not sure how to deal with the mrb_state. It's
not multithread safe, so I can't just use the mrb_state the Renderer
pointer points to, unless I only have a single thread, but seems like a
bad solution.

Instead, I can just spin up an entirely new VM, but configured the same
way. As long as I don't store any global state that I need access to, I
should be fine. I have no idea if that statement will hold or not, but
I'm gonna try it.