Refactor mruby code
I added some new utils for code reuse, RubyVM and Noise.

RubyVM will give us a new mrb_state* which we need for each Chunk's
thread when it generates the blocks. It applies all the Ruby and C++
extensions to the VM automatically, so it's ready for us to use.

Added a Noise class with a single static class method to get a perlin
noise value given an x,y,z.

Static class methods are easy to bind to ruby, which makes binding Noise
to Ruby straightforward.

I want to rename the PerlinNoise object in Ruby to match what it is in
C++.