Change some binding conventions
Anything in Ruby that mirrors 100% what's in C++ is going to be called
CSomething. So CNoise in ruby matches the C++ Noise class exactly, with
methods snake case versions of the C++ camelCase version.

I also make the C++ class a pure C++ wrapper, no mruby, and instead I
make the wrapper function inline with a lambda in the ruby_vm.cpp
spawnVM function.