Refactor block texture code
I wanted to pull some stuff into BlockEntry. Specifically, it keeps
track of the current block type now (based on the number of items in the
@block_entries array) and assigns one to each BlockEntry on create.

I also moved some of the texture logic in to the class. It's pretty
ugly code parsing that dsl, but anything dsl-parsing related should
be going in the class it's for, so I moved it there.