Fix all the segfaults
So, I've probably explained what's going on like 5 times in various
places, but basically:

When moving quickly across chunk boundaries, it was possible to end up
in a situation where a Chunk is initialized, then the world gen and mesh
thread is scheduled with a reference to that Chunk, *but* before that
thread could execute (the pool is full and executing on other chunks),
we cross another chunk boundary and that chunk we just initialized is
now out of range is is erased from the chunks map, then our thread
finally executes but the reference is now invalid! Segfault :(

The super simple solution was to use a shared_ptr to a Chunk instead of
the Chunk directly. This works with minimal changes, because we now can
call `chunks.erase` whenever, and if there is a thread with a shared_ptr
still to that same Chunk still, the object will hang around until the
thread is done with it.

The diligent reader will notice that it's possible we generate and mesh
a Chunk that is immediately destroyed. I'm not worried about this right
now, so I'm gonna ignore it.